Переносит состояние нажатой карточки в компонент
